# openapi.yaml
openapi: 3.1.0
info:
  title: Obor na dosah API
  version: 3.1.0
  summary: Výběr střední školy podle dojezdu a plánování oborů v Karlovarském kraji.
  description: |
    API slouží dvěma částem aplikace:

    - **Pro rodiny** – student zadá, odkud cestuje, a uvidí, které školy s jeho oborem stihne ráno dojet a jakým spojem.
    - **Pro kraj** – úředník vidí na mapě, odkud je obor v dosahu, které obory jsou přeplněné nebo nedostupné
      a co by změnilo přidání oboru na konkrétní školu.

    API je jen pro čtení. Mapové vrstvy jsou GeoJSON (WGS 84, EPSG:4326).
    Jediný endpoint, který volá OpenTripPlanner živě, je `/student/trasa`; ostatní čtou
    předpočítanou matici dojezdů ZSJ × škola.

    **Klíčové definice**

    | Pojem | Definice |
    | --- | --- |
    | `deti` | Odhad jednoho ročníku nastupujícího na SŠ v roce 2026: věková skupina 10–14 let ze SLDB 2021 v ZSJ, děleno 5. |
    | `cas_min` | Nejkratší spoj ze ZSJ k nejbližší škole s oborem (bez oboru k jakékoli SŠ), včetně chůze a přestupů. U obcí a ORP průměr ZSJ vážený počtem obyvatel. |
    | V dosahu | `cas_min` ≤ `max_min` (výchozí 120). |
    | `pasmo` | `do30`, `30_45`, `45_60`, `nad60` (60 až `max_min`), `mimo_dosah`, `bez_spojeni`. |
    | Děti bez oboru | Součet `deti` v ZSJ, kde obor v dosahu není. |
    | `prihlasky_na_misto` | Přihlášky / kapacita (přijímací řízení 2026, 1. kolo). |
    | `index_pretlaku` | `prihlasky_na_misto` / průměr kraje (2,66). 1,0 = průměr. |
...

    **Konvence**
...
servers:
...
paths:
  /skoly:
    get:
      tags: [Společné]
      operationId: listSkoly
      summary: Školy jako body na mapě
      description: Bez filtru vrátí všech 34 škol; 4 školy bez nabídky mají `pocet_nabidek` 0.
      parameters:
        - $ref: '#/components/parameters/OborFiltr'
        - $ref: '#/components/parameters/Stupen'
        - $ref: '#/components/parameters/Forma'
      responses:
        '500': { $ref: '#/components/responses/InterniChyba' }
        '503': { $ref: '#/components/responses/DatabazeNedostupna' }
        '200':
          description: GeoJSON bodů škol.
          content:
            application/geo+json:
              schema: { $ref: '#/components/schemas/SkolyFeatureCollection' }
        '422': { $ref: '#/components/responses/NeplatnyParametr' }
  /skoly/{redizo}:
    get:
      tags: [Společné]
      operationId: getSkola
      summary: Detail školy, její nabídka a spádová oblast
      parameters:
        - name: redizo
          in: path
          required: true
          schema: { $ref: '#/components/schemas/Redizo' }
          example: '600009084'
        - $ref: '#/components/parameters/MaxMin'
        - $ref: '#/components/parameters/Scenar'
      responses:
        '500': { $ref: '#/components/responses/InterniChyba' }
        '503': { $ref: '#/components/responses/DatabazeNedostupna' }
        '200':
          description: Detail školy.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SkolaDetail' }
        '404': { $ref: '#/components/responses/Nenalezeno' }
        '422': { $ref: '#/components/responses/NeplatnyParametr' }
...
  /zsj:
    get:
      tags: [Pro kraj]
      operationId: getZsj
      summary: Mapa – odkud je obor v dosahu a za kolik minut
      description: |
        Doba dojezdu k nejbližší škole, která obor učí (bez `obor` k nejbližší jakékoli SŠ),
        pro plochy na zvolené úrovni:

        | `uroven` | Plochy |
        | --- | --- |
        | `orp` | 7 obvodů pojmenovaných podle největších měst (Aš, Cheb, Karlovy Vary, Kraslice, Mariánské Lázně, Ostrov, Sokolov) |
        | `obec` | obce kraje (výchozí) |
        | `zsj` | 839 základních sídelních jednotek |

        U `orp` a `obec` je `cas_min` průměr ZSJ vážený počtem obyvatel (ZSJ bez spojení se nepočítají).
        `meta.v_limitu` počítá plochy do 30, 45, 60 a `max_min` minut.
      parameters:
        - $ref: '#/components/parameters/OborFiltr'
        - $ref: '#/components/parameters/Forma'
        - $ref: '#/components/parameters/Uroven'
        - $ref: '#/components/parameters/MaxMin'
        - $ref: '#/components/parameters/Scenar'
      responses:
        '501': { $ref: '#/components/responses/Neimplementovano' }
        '200':
          description: GeoJSON ploch s dobou dojezdu a pásmem.
          content:
            application/geo+json:
              schema: { $ref: '#/components/schemas/MapaDosahu' }
        '422': { $ref: '#/components/responses/NeplatnyParametr' }
  /obory:
    get:
      tags: [Pro kraj]
      operationId: listObory
      summary: Obory s bilancí – seznam pro výběr i tabulka pro úředníky
      parameters:
        - $ref: '#/components/parameters/Stupen'
        - $ref: '#/components/parameters/Forma'
        - $ref: '#/components/parameters/MaxMin'
        - $ref: '#/components/parameters/Scenar'
        - name: signal
          in: query
          description: Jen obory s daným signálem; více hodnot oddělit čárkou.
          schema: { type: string, example: 'pretlak,spatna_dostupnost' }
        - name: razeni
          in: query
          description: Pole pro řazení; mínus = sestupně.
          schema:
            type: string
            enum: [nazev, index_pretlaku, -index_pretlaku, deti_bez_oboru, -deti_bez_oboru,
                   mist_na_100_deti, -mist_na_100_deti, volna_mista_na_misto, -volna_mista_na_misto]
            default: nazev
      responses:
        '501': { $ref: '#/components/responses/Neimplementovano' }
        '200':
          description: Řádek za každý obor.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/BilanceOboru' }
                  meta: { $ref: '#/components/schemas/MetaBilance' }
        '422': { $ref: '#/components/responses/NeplatnyParametr' }
...
components:
  parameters:
...
    MaxMin:
      name: max_min
      in: query
      description: Nejdelší přijatelná doba dojezdu v minutách.
      schema: { type: integer, minimum: 10, maximum: 180, default: 120 }
    Scenar:
      name: scenar
      in: query
      description: |
        Scénář dojezdu („Příchod do školy“). Hodnoty odpovídají předpočítaným scénářům v tabulce `SCENAR`;
        dnes jen `rano` = příjezd 7:00–8:00 v běžný školní den.
      schema: { type: string, enum: [rano], default: rano }
    Uroven:
      name: uroven
      in: query
      description: Velikost ploch mapy – `orp` (7 obvodů největších měst), `obec`, `zsj`.
      schema: { type: string, enum: [orp, obec, zsj], default: obec }
    OborFiltr:
      name: obor
      in: query
      description: Kód oboru KKOV (v query není třeba lomítko kódovat).
      schema: { $ref: '#/components/schemas/KodOboru' }
      example: 18-20-M/01
    Stupen:
      name: stupen
      in: query
      description: Stupeň vzdělání (písmeno v kódu KKOV); více hodnot oddělit čárkou.
      schema: { type: string, pattern: '^[CEHKLM](,[CEHKLM])*$' }
      example: M
    Forma:
      name: forma
      in: query
      description: Forma vzdělávání.
      schema: { type: string, enum: [den, dal], default: den }
  responses:
    InterniChyba:
      description: Nepodařilo se načíst data nebo uložená data neodpovídají požadovanému formátu.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Chyba' }
    DatabazeNedostupna:
      description: Databázové připojení je dočasně nedostupné.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Chyba' }
    Neimplementovano:
      description: Endpoint je zatím pouze stub; jeho aplikační logika není implementovaná.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Chyba' }
    Nenalezeno:
      description: Záznam neexistuje.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Chyba' }
          example:
            error: { kod: nenalezeno, zprava: 'Škola 600000000 neexistuje.', pole: redizo }
    NeplatnyParametr:
      description: Neplatný parametr (neznámý obor nebo scénář, bod mimo kraj…).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Chyba' }
          example:
            error: { kod: mimo_kraj, zprava: 'Bod neleží v Karlovarském kraji.', pole: lat }
  schemas:
...
    KodObce: { type: string, pattern: '^\d{6}$', example: '554961' }
    KodPlochy: { type: string, pattern: '^\d{4,6}$', description: Kód ZSJ, obce nebo ORP podle `uroven`. }
    Redizo: { type: string, pattern: '^\d{9}$', example: '600009084' }
    KodOboru: { type: string, pattern: '^\d{2}-\d{2}-[A-Z]/\d{2}$', example: 18-20-M/01 }
    Stupen:
      type: string
      enum: [C, E, H, K, L, M]
      description: C praktická škola, E praktické obory, H výuční list, K gymnázium, L maturita s praxí / nástavba, M maturita.
    Signal:
      type: string
      enum: [pretlak, nizky_zajem, spatna_dostupnost, poptavka_trhu]
    Pasmo:
      type: string
      enum: [do30, '30_45', '45_60', nad60, mimo_dosah, bez_spojeni]
...
    Chyba:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [kod, zprava]
          properties:
            kod: { type: string }
            zprava: { type: string }
            pole: { type: [string, 'null'] }
    MetaDosah:
      type: object
      required: [scenar, max_min]
      properties:
        scenar: { type: string, example: rano }
        max_min: { type: integer, example: 120 }
    MetaBilance:
      allOf:
        - $ref: '#/components/schemas/MetaDosah'
        - type: object
          properties:
            deti_celkem: { type: integer }
            prumer_prihlasek_na_misto: { type: number, example: 2.66 }
            prijimaci_rizeni:
              type: object
              properties:
                rok: { type: integer, example: 2026 }
                kolo: { type: integer, example: 1 }
            prahy:
              type: object
              properties:
                pretlak: { type: number, example: 1.5 }
                nizky_zajem: { type: number, example: 0.5 }
                spatna_dostupnost: { type: number, example: 50 }
                poptavka_trhu: { type: number, example: 2 }
    VLimitu:
      type: object
      required: [limit_min, jednotek]
      properties:
        limit_min: { type: integer, example: 45 }
        jednotek: { type: integer, example: 46 }
...
    Nabidka:
      type: object
      required: [kod_oboru, nazev_oboru, forma, kapacita, prihlasky]
      properties:
        kod_oboru: { $ref: '#/components/schemas/KodOboru' }
        nazev_oboru: { type: string }
        zamereni: { type: [string, 'null'] }
        stupen: { $ref: '#/components/schemas/Stupen' }
        forma: { type: string, enum: [den, dal] }
        delka_let: { type: integer, example: 3 }
        kapacita: { type: integer, example: 30 }
        prihlasky: { type: integer, example: 131 }
        prihlasky_na_misto: { type: [number, 'null'], example: 4.37 }
        index_pretlaku: { type: [number, 'null'], example: 1.64 }
    GeoPoint:
      type: object
      required: [type, coordinates]
      properties:
        type: { type: string, const: Point }
        coordinates: { type: array, minItems: 2, maxItems: 2, items: { type: number }, example: [12.95214, 50.30477] }
    GeoPlocha:
      type: object
      required: [type, coordinates]
      properties:
        type: { type: string, enum: [Polygon, MultiPolygon] }
        coordinates: { type: array, items: {} }
...
    SkolyFeatureCollection:
      type: object
      required: [type, features]
      properties:
        type: { type: string, const: FeatureCollection }
        features:
          type: array
          items:
            type: object
            required: [type, geometry, properties]
            properties:
              type: { type: string, const: Feature }
              geometry: { $ref: '#/components/schemas/GeoPoint' }
              properties:
                type: object
                required: [redizo, nazev, pocet_nabidek]
                properties:
                  redizo: { $ref: '#/components/schemas/Redizo' }
                  nazev: { type: string }
                  obec: { type: string, example: Ostrov }
                  pocet_nabidek: { type: integer, example: 8 }
                  kapacita: { type: integer, example: 240 }
                  prihlasky: { type: integer, example: 724 }
                  index_pretlaku: { type: [number, 'null'], example: 1.14 }
    SkolaDetail:
      type: object
      required: [redizo, nazev, lat, lon, nabidky, spadovost]
      properties:
        redizo: { $ref: '#/components/schemas/Redizo' }
        nazev: { type: string }
        adresa: { type: string, example: 'Klínovecká 1197, 36301 Ostrov' }
        web: { type: [string, 'null'], format: uri }
        lat: { type: number }
        lon: { type: number }
        nabidky:
          type: array
          items: { $ref: '#/components/schemas/Nabidka' }
        spadovost:
          type: object
          description: Odkud se do školy dá dojet do `max_min`.
          properties:
            deti_v_dosahu: { type: integer }
            obce:
              type: array
              items:
                type: object
                properties:
                  kod_obce: { $ref: '#/components/schemas/KodObce' }
                  nazev: { type: string }
                  deti: { type: integer }
                  nejkratsi_cas_min: { type: integer }
        meta: { $ref: '#/components/schemas/MetaDosah' }
...
    PlochaDosahu:
      type: object
      required: [kod, nazev, uroven, pasmo, v_dosahu, deti]
      properties:
        kod: { $ref: '#/components/schemas/KodPlochy' }
        nazev: { type: string, example: Karlovy Vary }
        uroven: { type: string, enum: [orp, obec, zsj] }
        cas_min: { type: [integer, 'null'], example: 24, description: null = bez spojení. }
        pasmo: { $ref: '#/components/schemas/Pasmo' }
        v_dosahu: { type: boolean }
        deti: { type: integer, example: 412 }
        deti_v_dosahu: { type: integer, example: 412 }
        podil_deti_v_dosahu: { type: number, example: 100.0, description: '%' }
        nejblizsi_redizo: { type: [string, 'null'], example: '600009084' }
    MapaDosahu:
      type: object
      required: [type, features, meta]
      properties:
        type: { type: string, const: FeatureCollection }
        features:
          type: array
          items:
            type: object
            required: [type, geometry, properties]
            properties:
              type: { type: string, const: Feature }
              geometry: { $ref: '#/components/schemas/GeoPlocha' }
              properties: { $ref: '#/components/schemas/PlochaDosahu' }
        meta:
          allOf:
            - $ref: '#/components/schemas/MetaDosah'
            - type: object
              properties:
                obor: { type: [string, 'null'] }
                forma: { type: string }
                uroven: { type: string, enum: [orp, obec, zsj] }
                jednotek_celkem: { type: integer, example: 100 }
                v_limitu:
                  type: array
                  items: { $ref: '#/components/schemas/VLimitu' }
    BilanceOboru:
      type: object
      required: [kod, nazev, stupen, pocet_skol, kapacita, prihlasky, deti_v_dosahu, deti_bez_oboru, signaly]
      properties:
        kod: { $ref: '#/components/schemas/KodOboru' }
        nazev: { type: string, example: Informační technologie }
        stupen: { $ref: '#/components/schemas/Stupen' }
        pocet_skol: { type: integer, example: 3 }
        kapacita: { type: integer, example: 90 }
        prihlasky: { type: integer, example: 248 }
        prihlasky_na_misto: { type: [number, 'null'], example: 2.76 }
        index_pretlaku: { type: [number, 'null'], example: 1.04 }
        deti_v_dosahu: { type: integer }
        deti_bez_oboru: { type: integer }
        podil_deti_v_dosahu: { type: number, description: '%' }
        mist_na_100_deti: { type: [number, 'null'] }
        zamestnavatelu: { type: [integer, 'null'], description: Různá IČO s místy v souvisejících profesích; null = obor nemá mapování v NSP. }
        volna_mista: { type: [integer, 'null'], description: null = obor nemá mapování v NSP (nikdy 0 místo null). }
        volna_mista_na_misto: { type: [number, 'null'] }
        signaly:
          type: array
          items: { $ref: '#/components/schemas/Signal' }
...